+++ xbmc/cores/dvdplayer/DVDPlayerAudio.cpp
@@ -433,7 +433,7 @@ int CDVDPlayerAudio::DecodeFrame(DVDAudioFrame &audioframe)
           m_dvdAudio.Flush();
         m_dvdAudio.Pause();
       }
-      m_speed = speed;
+      m_speed = (int)speed;
     }
     else if (pMsg->IsType(CDVDMsg::AUDIO_SILENCE))
     {
